requirements.txt
python-gammu
flask
flask-restful
Flask-HTTPAuth
pyopenssl